requirements.txt
absl-py==2.0.0
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.4.0
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.4
attrs==23.1.0
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfscrape==2.1.1
charset-normalizer==3.2.0
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
comm==0.1.4
constantly==15.1.0
contourpy==1.1.1
cryptography==41.0.4
cssselect==1.2.0
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.5
docopt==0.6.2
docutils==0.20.1
easygui==0.98.3
EasyProcess==1.1
entrypoint2==1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
face-recognition-models==0.3.0
fastjsonschema==2.18.0
filelock==3.12.4
firebase==4.0.1
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozendict==2.3.8
gast==0.4.0
google==3.0.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-search==1.1.1
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
jax==0.4.16
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keras==2.13.1
kiwisolver==1.4.5
langdetect==1.0.9
libclang==16.0.6
lockfile==0.12.2
luigi==3.3.0
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.1
ml-dtypes==0.3.0
MouseInfo==0.1.3
mss==9.0.1
multitasking==0.0.11
mysql-connector-python==8.1.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
nltk==3.8.1
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandas-datareader==0.10.0
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
patsy==0.5.3
phonenumbers==8.13.20
pickleshare==0.7.5
Pillow==10.0.1
pipreqs==0.4.13
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
Protego==0.3.0
protobuf==4.21.12
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.19.0
PyDispatcher==2.0.7
PyGetWindow==0.0.9
Pygments==2.16.1
pygrametl==2.8
PyMsgBox==1.0.9
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyRect==0.2.0
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
queuelib==1.6.2
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
schedule==1.2.0
scikit-learn==1.3.0
scipy==1.11.2
Scrapy==2.11.0
seaborn==0.12.2
selenium==4.12.0
Send2Trash==1.8.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
Theano==1.0.5
threadpoolctl==3.2.0
tinycss2==1.2.1
tldextract==3.6.0
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
trio==0.22.2
trio-websocket==0.10.4
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
vboxapi==1.0
w3lib==2.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.3.7
wordcloud==1.9.2
wrapt==1.15.0
wsproto==1.2.0
yarg==0.1.9
yfinance==0.2.28
youtube-dl==2021.12.17
zope.interface==6.0
